+++ rx-location-manager-altitude/src/main/kotlin/net/samystudio/rxlocationmanager/altitude/RxLocationManagerAltitude.kt
@@ -12,11 +12,11 @@ import android.location.Location
 import android.os.Handler
 import androidx.annotation.RequiresPermission
 import androidx.annotation.VisibleForTesting
-import io.reactivex.Observable
-import io.reactivex.Observer
-import io.reactivex.Single
-import io.reactivex.disposables.Disposable
-import io.reactivex.functions.BiFunction
+import io.reactivex.rxjava3.core.Observable
+import io.reactivex.rxjava3.core.Observer
+import io.reactivex.rxjava3.core.Single
+import io.reactivex.rxjava3.disposables.Disposable
+import io.reactivex.rxjava3.functions.BiFunction
 import net.samystudio.rxlocationmanager.ContextProvider
 import net.samystudio.rxlocationmanager.RxLocationManager
 import net.samystudio.rxlocationmanager.nmea.GGA