+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.5.0-SNAPSHOT
+0.5.0-SNAPSHOT (rxjava3)
 ------------------
 - Add handler/looper arguments when possible to match Android API arguments.
 - Add getLastKnownLocation as Maybe observable.
@@ -11,6 +11,7 @@
 - Fix GGA quality enum list.
 - Change EnumValidator now uses CharArray insteadof array of char.
 - Rename observeGnssStatusChanged to observeGnssStatusOnChanged for consistency.
+- Migrate to rxjava3.
 
 0.3.0 (2018-11-07)
 ------------------